requirements.txt
numpy
pandas
Flask
scikit-learn
https://download.pytorch.org/whl/cpu/torch-1.7.0%2Bcpu-cp36-cp36m-linux_x86_64.whl
https://download.pytorch.org/whl/cpu/torchvision-0.8.1%2Bcpu-cp36-cp36m-linux_x86_64.whl
requests
Pillow
gunicorn == 20.0.4

Flask==2.0.1
Flask-Login==0.5.0
Flask-WTF==1.0.0
WTForms==3.0.1
numpy==1.21.2
pandas==1.3.3
Pillow==8.4.0
requests==2.26.0
reportlab==3.6.1